Move `ShapeContext1980` to feature types header
