+++ common/src/feature_types.cpp
@@ -66,4 +66,14 @@ namespace pcl
        << p.z_axis[0] << " " << p.z_axis[1] << " " << p.z_axis[2] << ")";
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const ShapeContext1980& p)
+  {
+    for (int i = 0; i < 9; ++i)
+    os << (i == 0 ? "(" : "") << p.rf[i] << (i < 8 ? ", " : ")");
+    for (std::size_t i = 0; i < 1980; ++i)
+      os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 1979 ? ", " : ")");
+    return (os);
+  }
 } // namespace pcl
