+++ common/src/point_types.cpp
@@ -279,16 +279,6 @@ namespace pcl
   }
 
   std::ostream& 
-  operator << (std::ostream& os, const ShapeContext1980& p)
-  {
-    for (int i = 0; i < 9; ++i)
-    os << (i == 0 ? "(" : "") << p.rf[i] << (i < 8 ? ", " : ")");
-    for (std::size_t i = 0; i < 1980; ++i)
-      os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 1979 ? ", " : ")");
-    return (os);
-  }
-
-  std::ostream& 
   operator << (std::ostream& os, const UniqueShapeContext1960& p)
   {
     for (int i = 0; i < 9; ++i)
