+++ features/include/pcl/features/3dsc.h
@@ -42,7 +42,7 @@
 
 #include <random>
 
-#include <pcl/point_types.h>
+#include <pcl/feature_types.h>
 #include <pcl/features/feature.h>
 
 namespace pcl
